Port large stripe removal from TomoPy to CuPy
Similar to other ports to CuPy, the residual is non-zero between the
CuPy version and the CPU TomoPy version, but small.

Notably, the result of both implementations when using default values
for all parameters produces projection images with noisy artifacts. It
is primarily the regions of noise which have the greatest difference in
the residual.

Currently, the reason for these artifacts is unknown, but one
possibility is that this particular stripe removal method is not suited
for the test data (ie, possibly no "large" stripes exist in the test
data).