Port dead stripe removal from TomoPy to CuPy
Similar comments here as the large stripe removal port in 8af3aa2
regarding the residual between the TomoPy/CPU implementation and the
CuPy/GPU implementation.

Something else worth noting is that the TomoPy version of this method
uses a 2D interpolation function from scipy
https://docs.scipy.org/doc/scipy/reference/generated/scipy.interpolate.interp2d.html.
The current latest version of CuPy (v11.5.0) does not have an analogue
of this function, but the pre-release version v.12.0.0b3 does, it has an
n-dimensional interpolation function that serves well as a GPU
replacement of the scipy `interp2d()` function
https://github.com/cupy/cupy/releases/tag/v12.0.0b3.

The required function `interpn()` in the CuPy codebase and its dependent
functions
https://github.com/cupy/cupy/blob/1a9c91411fe6297ea81c6ad103cd8211134f6446/cupyx/scipy/interpolate/_rgi.py
have been copied into the httomolib codebase for the moment, as it was
the simplest way to incorporate the required interpolation function.