Port phase retrieval from TomoPy to CuPy
Similar to other ports from TomoPy to CuPy, the residual between the two
implementations is non-zero, but small. In this case, the residual is
within a tolerance of 1e-7.

Something noteworthy in the port is that the TomoPy implementation makes
use of `ufunc.outer` to perform the operation `np.add.outer()`
https://numpy.org/doc/stable/reference/generated/numpy.ufunc.outer.html
in the `_reciprocal_grid()` function.

The current latest version of CuPy (v11.5.0) does not yet support
`ufunc.outer` (currently planned for v12, see
https://github.com/cupy/cupy/issues/6866#issue-1302905190, as well as
https://github.com/cupy/cupy/issues/7082 and
https://github.com/cupy/cupy/pull/7049). The equivalent result has been
generated explicitly using nested loops, but in the future when CuPy
makes this available in a release, the code should then be updated
accordingly.