Remove python version constraint in `meta.yaml`
It appeared that providing the constraint `>=3.9` for `python` in the
`meta.yaml` was preventing the `--python` flag for `conda build` from
working correctly (for example, passing in `--python 3.10` was failing
to install `python=3.10.x` in the build environment, and was instead
installing `python=3.11.x`). Therefore, that version constraint has been
removed from `meta.yaml`.